chore: Start updating AppBarButton
